Split test-utils fixture.ts into focused modules
Break up the ~374-line fixture.ts whose responsibilities had drifted
(command execution, git env isolation, version pinning, tarball
matching, workspace discovery, and two fixture constructors) into
focused modules:

- lib/command.ts — CommandResult, runCommand, exec, createGitEnv
- lib/workspace.ts — PackageJson schema, workspace root/index memo,
  resolveWorkspaceDeps
- lib/tarball.ts — matchTarball, pinned, tarball memo, resolveTarballs
- isolated-fixture.ts — createIsolatedFixture + types
- project-fixture.ts — createProjectFixture, extendWithFixture + types
- fixture.ts — thin barrel re-exporting the public surface

Public API and behavior unchanged. Memo caches split cleanly: workspace
root/index live in workspace.ts, tarball cache in tarball.ts — no
cross-module state.

Closes #55
